+++ lyra2/lyra2RE.cu
@@ -63,9 +63,9 @@ extern "C" int scanhash_lyra2(int thr_id, uint32_t *pdata,
 	unsigned long *hashes_done)
 {
 	const uint32_t first_nonce = pdata[19];
-	int intensity = (device_sm[device_map[thr_id]] > 500) ? 256 * 256 * 25 : 256 * 256 * 14;
-	int throughput = (int) device_intensity(thr_id, __func__, 1U << intensity); // 18=256*256*4;
-	throughput = min(throughput, (int)(max_nonce - first_nonce));
+	unsigned int intensity = (device_sm[device_map[thr_id]] > 500) ? 256 * 256 * 25 : 256 * 256 * 14;
+	uint32_t throughput = device_intensity(thr_id, __func__, 1U << intensity); // 18=256*256*4;
+	throughput = min(throughput, (max_nonce - first_nonce));
 
 	if (opt_benchmark)
 		((uint32_t*)ptarget)[7] = 0x000f;
