+++ quark/quarkcoin.cu
@@ -139,7 +139,7 @@ extern "C" int scanhash_quark(int thr_id, uint32_t *pdata,
 {
 	const uint32_t first_nonce = pdata[19];
 
-	int throughput = (int) device_intensity(thr_id, __func__, 1 << 20); // 256*4096
+	uint32_t throughput = device_intensity(thr_id, __func__, 1 << 20); // 256*4096
 	throughput = min(throughput, max_nonce - first_nonce);
 
 	if (opt_benchmark)
