+++ x13/x13.cu
@@ -156,10 +156,10 @@ extern "C" int scanhash_x13(int thr_id, uint32_t *pdata,
 	const uint32_t first_nonce = pdata[19];
 	static bool init[MAX_GPUS] = { 0 };
 	uint32_t endiandata[20];
-	int intensity = 19; // (device_sm[device_map[thr_id]] > 500 && !is_windows()) ? 20 : 19;
-	int throughput = (int) device_intensity(thr_id, __func__, 1 << intensity); // 19=256*256*8;
+	unsigned int intensity = 19; // (device_sm[device_map[thr_id]] > 500 && !is_windows()) ? 20 : 19;
+	uint32_t throughput = device_intensity(thr_id, __func__, 1 << intensity); // 19=256*256*8;
 
-	throughput = min(throughput, (int)(max_nonce - first_nonce));
+	throughput = min(throughput, (max_nonce - first_nonce));
 
 	if (opt_benchmark)
 		((uint32_t*)ptarget)[7] = 0xf;
